# openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "beton",
    "description": "REST-API von beton. Fehler sind Problem-Objekte nach RFC 9457.",
    "contact": {
      "name": "Ingo Fahrentholz"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:7420",
      "description": "Lokaler Daemon (Standard-Port)"
    }
  ],
  "paths": {
    "/auth/local/redeem": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Einmal-Code einlösen: setzt das Session-Cookie und leitet ohne Code weiter (AUTH-004).",
        "operationId": "redeem_login_code",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "next",
            "in": "query",
            "description": "Relativer Pfad für die Weiterleitung (Default `/`).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Weiterleitung mit Set-Cookie"
          },
          "400": {
            "description": "Ungültige Anfrage",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Host, Origin oder Berechtigung abgelehnt",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Interner Fehler (mit trace_id)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/v1/auth/local/codes": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Einmal-Code erzeugen (AUTH-004). Nur mit lokalem Token und nur über Loopback.",
        "operationId": "create_login_code",
        "responses": {
          "201": {
            "description": "Einmal-Code erzeugt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginCode"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Anfrage",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Host, Origin oder Berechtigung abgelehnt",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Interner Fehler (mit trace_id)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/hosts/{id}": {
      "get": {
        "tags": [
          "hosts"
        ],
        "summary": "Ein Host mit den unverändert gemeldeten Provider-Capabilities (RUN-001 AC3).",
        "operationId": "get_host",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Host-ID, lokal `hst_local`",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Host mit Providern",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Anfrage",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Host, Origin oder Berechtigung abgelehnt",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Host unbekannt",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Interner Fehler (mit trace_id)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/info": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Version, Schema- und Protokollversion des Servers und aktive Feature-Flags.",
        "operationId": "info",
        "responses": {
          "200": {
            "description": "Server-Informationen",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Anfrage",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Host, Origin oder Berechtigung abgelehnt",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Interner Fehler (mit trace_id)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Der angemeldete Benutzer und die Art der Anmeldung.",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Angemeldeter Benutzer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Anfrage",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Host, Origin oder Berechtigung abgelehnt",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Interner Fehler (mit trace_id)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/openapi.json": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Das OpenAPI-Dokument (API-001).",
        "operationId": "openapi_json",
        "responses": {
          "200": {
            "description": "OpenAPI 3.1",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "Ungültige Anfrage",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Host, Origin oder Berechtigung abgelehnt",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Interner Fehler (mit trace_id)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "type": "object",
        "description": "Ein Validierungsfehler mit JSON-Pointer auf das Feld.",
        "required": [
          "pointer",
          "detail"
        ],
        "properties": {
          "pointer": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "HostInfo": {
        "type": "object",
        "description": "Ein Host (lokal genau `hst_local`).",
        "required": [
          "id",
          "name",
          "providers"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "hst_local"
          },
          "name": {
            "type": "string"
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderInfo"
            }
          }
        }
      },
      "Info": {
        "type": "object",
        "description": "Server-Informationen.",
        "required": [
          "version",
          "schema_version",
          "protocol_version",
          "mode",
          "org_id",
          "features"
        ],
        "properties": {
          "version": {
            "type": "string"
          },
          "schema_version": {
            "type": "integer",
            "format": "int64"
          },
          "protocol_version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "mode": {
            "type": "string",
            "description": "`local` (M0) oder `server`."
          },
          "org_id": {
            "type": "string",
            "example": "org_local"
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Aktive Feature-Flags (UX-007); Funktionen hinter anderen Flags blendet die UI aus."
          }
        }
      },
...
      "Isolation": {
        "type": "string",
        "enum": [
          "process",
          "container",
          "pod"
        ]
      },
      "LoginCode": {
        "type": "object",
        "description": "Ein Einmal-Code für die Browser-Anmeldung.",
        "required": [
          "code",
          "redeem_url",
          "expires_in_s"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "redeem_url": {
            "type": "string",
            "description": "Fertige URL zum Öffnen im Browser."
          },
          "expires_in_s": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "Me": {
        "type": "object",
        "description": "Der angemeldete Benutzer.",
        "required": [
          "id",
          "org_id",
          "auth"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "usr_local"
          },
          "org_id": {
            "type": "string",
            "example": "org_local"
          },
          "auth": {
            "type": "string",
            "description": "`bearer` oder `cookie`."
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "Problem-Objekt nach RFC 9457.",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "trace_id"
        ],
        "properties": {
          "type": {
            "type": "string",
            "description": "`urn:beton:problem:<code>`"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "code": {
            "type": "string",
            "description": "Maschinenlesbarer Code; offener Katalog (neue Codes sind kein Breaking Change),\nReferenz: `docs/generated/problem-codes.md`.",
            "example": "not_found"
          },
          "trace_id": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "additionalProperties": false
      },
...
      "ProviderInfo": {
        "type": "object",
        "description": "Ein Provider eines Hosts mit seinen Capabilities (RUN-001).",
        "required": [
          "id",
          "capabilities"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "capabilities": {
            "$ref": "#/components/schemas/RunnerCapabilities"
          }
        }
      },
...
      "ResourceKind": {
        "type": "string",
        "enum": [
          "cpu",
          "memory",
          "pids",
          "disk",
          "timeout"
        ]
      },
      "RunnerCapabilities": {
        "type": "object",
        "description": "Was ein Provider kann (RUN-001). Der Server kennt Provider nur hierüber.",
        "required": [
          "isolation",
          "workspace_modes",
          "persistent_workspace",
          "interactive_exec",
          "snapshot",
          "resource_limits",
          "platforms"
        ],
        "properties": {
          "isolation": {
            "$ref": "#/components/schemas/Isolation"
          },
          "workspace_modes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceMode"
            }
          },
          "persistent_workspace": {
            "type": "boolean"
          },
          "interactive_exec": {
            "type": "boolean",
            "description": "PTY-Exec, z. B. für Device-Flow-Logins."
          },
          "snapshot": {
            "type": "boolean"
          },
          "resource_limits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceKind"
            }
          },
          "platforms": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "`os/arch` der erzeugten Runner, z. B. `linux/x86_64`."
          },
          "max_concurrent": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "WorkspaceMode": {
        "type": "string",
        "enum": [
          "host_path",
          "bind",
          "volume",
          "clone"
        ]
      }
    },
    "securitySchemes": {
      "browser_session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "beton_session"
      },
      "local_token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "local_token": []
    },
    {
      "browser_session": []
    }
  ]
}